+++ .gitignore
@@ -15,6 +15,7 @@ build
 
 # ide
 .vscode
+.idea
 
 npm-debug.log*
 yarn-debug.log*
