+++ src/shared/ui/checkbox/styles.module.css
@@ -12,7 +12,7 @@
   color: var(--colors-interface-black);
 }
 
-.checkbox-icon--wrapper {
+.checkboxIconWrapper {
   box-sizing: border-box;
   display: inline-flex;
   align-items: center;
@@ -26,38 +26,40 @@
   content: '';
   border: 2px solid var(--colors-interface-yorkshire-cloud);
   border-radius: 4px;
-
+  transition: var(--transition-hover);
   background-color: white;
 }
 
-.checkbox-icon {
+.checkboxIcon {
   position: absolute;
   width: 100%;
   height: 100%;
   visibility: hidden;
 }
 
-.checkbox:disabled+label > .checkbox-icon--wrapper {
+.checkbox:disabled+label > .checkboxIconWrapper {
   opacity: .4;
 }
 
-.checkbox:hover+label > .checkbox-icon--wrapper {
+.checkbox:hover+label > .checkboxIconWrapper {
   cursor: pointer;
   background-color: var(--colors-interface-primary);
   border-color: var(--colors-interface-primary);
   opacity: .4;
+  transition: var(--transition-hover);
 }
 
-.checkbox:checked+label > .checkbox-icon--wrapper {
+.checkbox:checked+label > .checkboxIconWrapper {
   cursor: pointer;
   background-color: var(--colors-interface-primary);
   border-color: var(--colors-interface-primary);
+  transition: var(--transition-hover);
 }
 
-.checkbox:hover+label > .checkbox-icon--wrapper > .checkbox-icon {
+.checkbox:hover+label > .checkboxIconWrapper > .checkboxIcon {
   visibility: visible;
 }
 
-.checkbox:checked+label > .checkbox-icon--wrapper > .checkbox-icon {
+.checkbox:checked+label > .checkboxIconWrapper > .checkboxIcon {
   visibility: visible;
 }
